refactor: missing refactor pieces
